% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interplot_default.R
\name{interplot.default}
\alias{interplot.default}
\title{Plot Conditional Coefficients in (Generalized) Linear Models with Interaction Terms}
\usage{
\method{interplot}{default}(m, var1, var2, plot = TRUE, point = FALSE,
  sims = 5000, xmin = NA, xmax = NA)
}
\arguments{
\item{m}{A model object including an interaction term, or, alternately, a data frame recording conditional coefficients.}

\item{var1}{The name (as a string) of the variable of interest in the interaction term; its conditional coefficient estimates will be plotted.}

\item{var2}{The name (as a string) of the other variable in the interaction term.}

\item{plot}{A logical value indicating whether the output is a plot or a dataframe including the conditional coefficient estimates of var1, their upper and lower bounds, and the corresponding values of var2.}

\item{point}{A logical value determining the format of plot. By default, the function produces a line plot when var2 takes on ten or more distinct values and a point (dot-and-whisker) plot otherwise; option TRUE forces a point plot.}

\item{sims}{Number of independent simulation draws used to calculate upper and lower bounds of coefficient estimates: lower values run faster; higher values produce smoother curves.}

\item{xmin}{A numerical value indicating the minimum value shown of x shown in the graph. Rarely used.}

\item{xmax}{A numerical value indicating the maximum value shown of x shown in the graph. Rarely used.}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
\code{interplot.default} is a method to calculate conditional coefficient estimates from the results of (generalized) linear regression models with interaction terms.
}
\details{
\code{interplot.default} is a S3 method from the \code{interplot}. It works on two classes of objects:
\itemize{
  \item Ordinary linear models (object class: \code{lm});
  \item Generalized linear models (object class: \code{glm}).
  }

The examples below illustrate how methods invoked by this generic deal with different type of objects.

Because the output function is based on \code{\link[ggplot2]{ggplot}}, any additional arguments and layers supported by \code{ggplot2} can be added with the \code{+}.
}

