% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interplot.R
\name{interplot}
\alias{interplot}
\title{INTERPLOT: Plot Interactive Effects}
\usage{
interplot(m, ...)
}
\arguments{
\item{m}{A R object storing the result of a regression including at least one interaction term.}

\item{var1}{The target variable in the interaction term of interest, whose coefficient changes is intended to plot.}

\item{var2}{The changing variable in the interaction term of interest, along with the changes of which the changes of the coefficient of the other variable is shown in the plot.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{labels}{A logical value to deside whether using the variable labels.}

\item{seed}{An arbitrary numeric value.}

\item{sims}{Number of independent simulation draws to create.}

\item{steps}{Desired length of the sequence. A non-negative number, which for seq and seq.int will be rounded up if fractional.}

\item{xmin}{a numerical value deciding the minimum value shown of x shown in the graph.}

\item{xmax}{a numerical value deciding the maximum value shown of x shown in the graph.}

\item{plot}{A logical value to deside the output is a plot or a list of the coefficient, upper and lower bound of var2.}
}
\description{
interplot is the basic plotting function in the interplot package. It is a convenient function for creating graphs describing how the coefficient of one variable in a two-way interaction term will change along with the changes of the other variable.
}
\examples{
# Create continuous, dummy, missing data, and group level variables.

group<-seq(1, 50, 1)
z<-rnorm(50, 1, 1)
u<-rnorm(50, 0, 3)
df<-data.frame(group=rep(group, 50), z=rep(z,50), u=rep(u,50))
df$x1<-rnorm(2500, 3, 1)+0.1*(group-1)
df$d<-rbinom(2500, 1, 0.2)

df$x2<-sample(1:10, 2500, T)
df$x2.miss <- df$x2
df$x2.miss[df$group < 5 & df$x1 < 4] <- NA
df$e<-rnorm(2500, 0, 2)
df$y<-2 - df$x1 + 0.3*df$x2 + 0.5*df$z + df$u + df$e

# Apply the interplot to different regressions
library(Interplot)

## 1. OLS
m1<-lm(y~x1+x2+d+z+x1:z, data = df)
interplot(m1, "x1", "z")

## 2. Logit with two interaction terms (the second term is of interest)
m2<-glm(d~y+x1+x2+z+x1:z+y:z, family=binomial(link="logit"), data = df)
interplot(m2, "y", "z")

## 3. Multilevel
m3<-lmer(y~x1+x2+d+z+x1:z+(1|group), data = df)
interplot(m3, "x1","z")

## 4. Multiple Imputed Data
library(Amelia)
m.imp <- amelia(df, idvars = c("y", "x2"))

m4 <- lapply(m.imp$imputations, function(i) lm(y ~ x1 + x2.miss + d + x2.miss*z, data = i))
interplot(m4, "x2.miss","z")
}

