% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interplot_mi.R
\name{interplot.lmmi}
\alias{interplot.lmmi}
\title{Graph Interaction Effects in Regular Models with Imputed Data}
\usage{
\method{interplot}{lmmi}(m, var1, var2, xlab = NULL, ylab = NULL,
  seed = 324, sims = 1000, steps = 100, xmin = NA, xmax = NA,
  labels = NULL, plot = TRUE)
}
\arguments{
\item{m}{a model object including the interaction of interest.}

\item{var1}{The target variable in the interaction term of interest, whose coefficient changes is intended to plot.}

\item{var2}{The changing variable in the interaction term of interest, along with the changes of which the changes of the coefficient of the other variable is shown in the plot.}

\item{xlab}{A character variable to set the title for the x axis.}

\item{ylab}{A character variable to set the title for the y axis.}

\item{seed}{An arbitrary numeric value. The default value is 324.}

\item{sims}{Number of independent simulation draws to create.}

\item{steps}{Desired length of the sequence. A non-negative number, which for seq and seq.int will be rounded up if fractional.}

\item{xmin}{a numerical value deciding the minimum value shown of x shown in the graph.}

\item{xmax}{a numerical value deciding the maximum value shown of x shown in the graph.}

\item{labels}{A logical value to deside whether using the variable labels.}

\item{plot}{A logical value to deside the output is a plot or a list of the coefficient, upper and lower bound of var2.}

\item{...}{additional arguments affecting the output of the plot.}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
\code{interplot.default} is a method to graph interaction effects from the results of linear and general linear regression models based on imputed data.
}
\details{
\code{interplot} is a S3 method from the \code{interplot}. It can visualize the changes in the coefficient of one term in a two-way interaction conditioned by the other term. This function can work on interactions from results in the class of \code{list}.

Because the output function is based on \code{\link[ggplot2]{ggplot}}, any additional arguments and layers supported by \code{ggplot2} can be added with the \code{+}.
}
\examples{
# Create continuous, dummy, missing data, and group level variables.

group<-seq(1, 50, 1)
z<-rnorm(50, 1, 1)
u<-rnorm(50, 0, 3)
df<-data.frame(group=rep(group, 50), z=rep(z,50), u=rep(u,50))
df$x1<-rnorm(2500, 3, 1)+0.1*(group-1)
df$d<-rbinom(2500, 1, 0.2)

df$x2<-sample(1:10, 2500, TRUE)
df$x2.miss <- df$x2
df$x2.miss[df$group < 5 & df$x1 < 4] <- NA
df$e<-rnorm(2500, 0, 2)
df$y<-2 - df$x1 + 0.3*df$x2 + 0.5*df$z + df$u + df$e

# Apply the interplot to different regressions
library(Interplot)
library(Amelia)
m.imp <- amelia(df, idvars = c("y", "x2"))

m1 <- lapply(m.imp$imputations, function(i) lm(y ~ x1 + x2.miss + d + x2.miss*z, data = i))
interplot(m1, "x2.miss","z")
}

