% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interplot_plot.R
\name{interplot.plot}
\alias{interplot.plot}
\title{Visualize Numeric Statistics of An Interaction}
\usage{
\method{interplot}{plot}(m, ylab = NULL, xlab = NULL, ...)
}
\arguments{
\item{m}{a data.frame recording the statistics of the changes caused by the interaction. The data.frame should includes four columns:
\itemize{
   \item fake: the sequence of \code{var1} (the item whose effect will be conditioned on in the interaction);
   \item coef1: the point estimates of the coefficient of \code{var1} at each break point.
   \item ub: the upper bound of the simulated 95\% CI.
   \item lb: the lower bound of the simulated 95\% CI.
}}

\item{ylab}{set the label of the y axis.}

\item{xlab}{set the label of the x axis.}

\item{...}{additional arguments from \code{\link[ggplot2]{ggplot}}.}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
Graph based on the data frame of statistics about the conditional effect of an interaciton.
}
\details{
\code{interplot.plot} is a S3 method from the \code{interplot}. It can graph the interaction effects based on a statistic data.frame recording the effects in a numeric way.

Because the output function is based on \code{\link[ggplot2]{ggplot}}, any additional arguments and layers supported by \code{ggplot2} can be added with the \code{+}.
}

