% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interplot.R
\name{interplot}
\alias{interplot}
\title{Plot Interactive Effects}
\usage{
interplot(m, var1, var2, xlab = NULL, ylab = NULL, seed = 324,
  sims = 1000, steps = NULL, xmin = NA, xmax = NA, labels = NULL,
  plot = TRUE, point = FALSE)
}
\arguments{
\item{m}{A model object including the interaction of interest, or a data.frame recording the statistics of the changes caused by the interaction.}

\item{var1}{A charecter value showing the target variable in the interaction term of interest, whose coefficient changes is intended to plot.}

\item{var2}{A charecter value showing the changing variable in the interaction term of interest, along with the changes of which the changes of the coefficient of the other variable is shown in the plot.}

\item{xlab}{A character variable to set the title for the x axis.}

\item{ylab}{A character variable to set the title for the y axis.}

\item{seed}{An arbitrary numeric value. The default value is 324.}

\item{sims}{Number of independent simulation draws to create.}

\item{steps}{Desired length of the sequence. A non-negative number, which for seq and seq.int will be rounded up if fractional. The default is 100 or the unique categories in the \code{var2} (when it is less than 100. Also see \code{\link{unique}}).}

\item{xmin}{A numerical value deciding the minimum value shown of x shown in the graph.}

\item{xmax}{A numerical value deciding the maximum value shown of x shown in the graph.}

\item{labels}{A logical value to deside whether using the variable labels.}

\item{plot}{A logical value to deside the output is a plot or a list of the coefficient, upper and lower bound of var2.}

\item{point}{A logical value determining the format of plot. The function produces a point plot when it is \code{TRUE}; otherwise, the fucntion produces a line plot. Both plots have 95\% confidential intervals. The default is \code{FAULSE}.}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
\code{interplot} is a generic function to produce plot of the conditional graph of one variable in an two-way interaction conditioned on the other variable from the results of various model fitting functions. The function invokes particular \code{methods} which depend on the \code{\link{class}} of the first argument.
}
\details{
\code{interplot} visualizes the changes in the coefficient of one term in a two-way interaction conditioned by the other term. In the current version, the function can work on interactions in four classes of objects:
\itemize{
  \item Ordinary linear models (object class: \code{lm});
  \item Generalized linear models (object class: \code{glm});
  \item Ordinary linear models with imputed data (object class: \code{list});
  \item Generalized linear models with imputed data (object class: \code{list}).
}

The examples below illustrate how methods invoked by this generic deal with different type of objects.

Because the output function is based on \code{\link[ggplot2]{ggplot}}, any additional arguments and layers supported by \code{ggplot2} can be added with the \code{+}.
}
\examples{
data(mtcars)
m_cyl <- lm(mpg ~ wt * cyl, data = mtcars)
library(interplot)

# Plot interactions with a continous conditioning variable
interplot(m = m_cyl, var1 = "cyl", var2 = "wt") +
ggtitle("The Conditional Effect of Weight \\n
on the Influence of Cylinder Number") +
theme(plot.title = element_text(face="italic"))


# Plot interactions with a categorical conditioning variable
interplot(m = m_cyl, var1 = "wt", var2 = "cyl")
}

