% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interplot.R
\name{interplot}
\alias{interplot}
\title{Plot Interactive Effects}
\usage{
interplot(m, ...)
}
\arguments{
\item{m}{a model object including the interaction of interest.}

\item{...}{additional arguments affecting the output of the plot.}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
\code{interplot} is a generic function to produce plot of the conditional graph of one variable in an two-way interaction conditioned on the other variable from the results of various model fitting functions. The function invokes particular \code{methods} which depend on the \code{\link{class}} of the first argument.
}
\details{
\code{interplot} visualizes the changes in the coefficient of one term in a two-way interaction conditioned by the other term. In the current version, the function can work on interactions in four classes of objects:
\itemize{
  \item Ordinary linear models (object class: \code{lm});
  \item Generalized linear models (object class: \code{glm});
  \item Ordinary linear models with imputed data (object class: \code{list});
  \item Generalized linear models with imputed data (object class: \code{list}).
}

The examples below illustrate how methods invoked by this generic deal with different type of objects.

Because the output function is based on \code{\link[ggplot2]{ggplot}}, any additional arguments and layers supported by \code{ggplot2} can be added with the \code{+}.
}
\examples{
data(mtcars)
m_cyl <- lm(mpg ~ wt * cyl, data = mtcars)
library(interplot)

# Plot interactions with a continous conditioning variable
interplot(m = m_cyl, var1 = "cyl", var2 = "wt") +
ggtitle("The Conditional Effect of Weight \\n
on the Influence of Cylinder Number") +
theme(plot.title = element_text(face="italic"))


# Plot interactions with a categorical conditioning variable
interplot(m = m_cyl, var1 = "wt", var2 = "cyl")
}

