% Generated by roxygen2 (4.0.1): do not edit by hand
\name{interplot}
\alias{interplot}
\title{INTERPLOT: Plot Interactive Effects}
\usage{
interplot(m, var1, var2, xlab = NULL, ylab = NULL, seed = 313,
  sims = 1000, steps = 100, plot = TRUE)
}
\arguments{
\item{m}{A R object storing the result of a regression including at least one interaction term.}

\item{var1}{The target variable in the interaction term of interest, whose coefficient changes is intended to plot.}

\item{var2}{The changing variable in the interaction term of interest, along with the changes of which the changes of the coefficient of the other variable is shown in the plot.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{seed}{An arbitrary numeric value.}

\item{sims}{Number of independent simulation draws to create.}

\item{steps}{Desired length of the sequence. A non-negative number, which for seq and seq.int will be rounded up if fractional.}

\item{plot}{A logical value to deside the output is a plot or a list of the coefficient, upper and lower bound of var2.}
}
\description{
interplot is the basic plotting function in the interplot package. It is a convenient function for creating graphs describing how the coefficient of one variable in a two-way interaction term will change along with the changes of the other variable.
}
\examples{
# Create continuous, dummy, and group level variables.
y<-rnorm(100, 1, 2)
x<-log(abs(y))
z<-y^2
d<-ifelse(y<1, 0, 1)
m<-ifelse(y<c(-3), 4,
          ifelse(y<0, 3,
                 ifelse(y<3, 2, 1)))

# Apply the interplot to different regressions
library(Interplot)

## 1. OLS
m1<-lm(y~x+d+z+x:z)
interplot(m1, "x", "z")

## 2. Logit with two interaction terms (the second term is of interest)
m2<-glm(d~y+x+z+x:z+y:z, family=binomial(link="logit"))
interplot(m2, "y", "z")

## 3. Multilevel
m3<-lmer(y~x+d+z+x:z+(1|m))
interplot(m3, "x","z")
}

